refactor: 공통 유틸리티 폴더 구조 개선
- utils 폴더를 commons/utils로 이동
